Refactor everything about Cucumber messages
Messages are no longer lingering in the runtime, but passed along to the
backend using tasks immediately. This is a prerequisite for the
following issues:

 - 810 [1]
 - 824 [2]
 - 944 [3]

This is not to say that these will be fixed shortly or even at all, but
at least now it will be remotely possible.

Furthermore, this removes any doubt regarding whether reports work in
interactive mode. The experimental flag simply turns out to be too
broken [4-5]. Interactive mode is now detected and reports turned off.

Lastly, detecting interactive mode is a somewhat broken feature [6]. The
workaround (using `isTextTerminal`) is only available in Cypress v10 and
beyond. Thus is Cypress v9 now unsupported.

[1] https://github.com/badeball/cypress-cucumber-preprocessor/issues/810
[2] https://github.com/badeball/cypress-cucumber-preprocessor/issues/824
[3] https://github.com/badeball/cypress-cucumber-preprocessor/issues/944
[4] https://github.com/cypress-io/cypress/issues/18955
[5] https://github.com/cypress-io/cypress/issues/26634
[6] https://github.com/cypress-io/cypress/issues/20789